requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

appnope==0.1.0
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
chart-studio==1.0.0
conda==4.3.16
cssselect==1.1.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
jieba3k==0.35.1
Jinja2==2.10
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
lxml==4.4.1
MarkupSafe==1.1.0
mistune==0.8.4
mutagen==1.42.0
nbconvert==5.6.0
nbformat==4.4.0
newspaper3k==0.2.8
ngrok==0.0.1
nltk==3.4.5
notebook==6.0.1
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
plotly==4.5.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycosat==0.6.3
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-file==1.4.3
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.4
splinter==0.11.0
statsmodels==0.11.1
terminado==0.8.2
testpath==0.4.2
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.6
virtualenv==16.7.8
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
